models/HDU/Hangzhou Dianzi University/Model/requirements.txt
numpy
Pillow
scikit-learn
matplotlib
pandas
seaborn

